requirements.txt

django-crispy-forms==1.7.2
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
